Move diff caclulation to tableau crate
