+++ rust/gitxetcore/src/diff/twb.rs
@@ -1,23 +1,27 @@
-use serde::{Deserialize, Serialize};
+use tableau_summary::twb::diff::{TwbDiffProcessor};
 use tableau_summary::twb::TwbSummary;
 use crate::diff::output::SummaryDiffData;
 use crate::diff::{DiffError, SummaryDiffProcessor};
 use crate::diff::output::SummaryDiffData::Twb;
 use crate::summaries::{FileSummary, SummaryType};
 
-/// Diff content for a Twb diff
-#[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
-pub struct TwbSummaryDiffContent {
-    pub before: Option<TwbSummary>,
-    pub after: Option<TwbSummary>,
-}
 
 /// Processes diffs of Twb Summaries, taking in [TwbSummary]s and producing a [TwbSummaryDiffContent]
 /// indicating the delta between them.
 ///
 /// Currently, this will just provide the two summaries (before + after). In the future, we can
 /// change this to become more intelligent (actually identify what changed between the workbooks).
-pub struct TwbSummaryDiffProcessor {}
+pub struct TwbSummaryDiffProcessor {
+    proc: TwbDiffProcessor
+}
+
+impl TwbSummaryDiffProcessor {
+    pub fn new() -> Self {
+        TwbSummaryDiffProcessor {
+            proc: TwbDiffProcessor {},
+        }
+    }
+}
 
 
 impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
@@ -37,17 +41,11 @@ impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
     }
 
     fn get_insert_diff(&self, summary: &TwbSummary) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(TwbSummaryDiffContent {
-            before: None,
-            after: Some(summary.clone()),
-        }))
+        Ok(Twb(self.proc.get_insert_diff(summary)?))
     }
 
     fn get_remove_diff(&self, summary: &TwbSummary) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(TwbSummaryDiffContent {
-            before: Some(summary.clone()),
-            after: None,
-        }))
+        Ok(Twb(self.proc.get_remove_diff(summary)?))
     }
 
     fn get_diff_impl(
@@ -55,9 +53,6 @@ impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
         before: &TwbSummary,
         after: &TwbSummary,
     ) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(TwbSummaryDiffContent {
-            before: Some(before.clone()),
-            after: Some(after.clone()),
-        }))
+        Ok(Twb(self.proc.get_diff_impl(before, after)?))
     }
 }
